Extend Places class to support executive attribute
This allows us to support places that have an executive, such as
countries. It's not required so places without an executive still work.

The MethodLength cop has been disabled on the data method because it's
over 10 lines. I think this is OK because the extra lines aren't
actually _doing_ anything, they're just making it easier to read setting
the attributes of the Division that's returned.