Support multiple legislatures in CountryCities and CountryDivisions
In Wikidata items can have multiple statements for properties such as
this one - Legislative Body For (P194). This change allows us to
show when this is the case for a these places.

There is still the issue that multiple results will appear for places
with multiple P194s. This will need to be addressed in subsequent
commits.