make a fresh Reek config
We inherited the existing one from `viewer-sinatra` and it's full of
exlusions that only apply there. Start with a fresh one here, with no
exclusions for now, other than no needing documentation (we can address
that issue separately).

This currently has various failures that we should look at on a case by
case basis.