Introduce a SelfAwareStruct
The results of SPARQL queries usually require a LabelledItem to
represent the primary item we're dealing with (e.g. the country, or
legislature), as well as other LabelledItems to represent associated
concepts. But if we treat all of those the same, we end up with the ugly
indirection of having to say `country.country.id` instead of just
`country.id`

So add a subclass of Struct that, when given an argment of `:me`, knows
how to call `id` and `name` on that.

This is a bit of a nasty hack that's pushing the concept of a Struct
right to its limit. We might want to look into pushing this up to a
full-blown class, but for now it's an experiment whilst we see if it
has sufficient value first.