Fix undefined method `id' exception on country pages
In cee631b we introduced auto-inflating LabelledItems that changed the
way the data that's passed to the view is structured. This resulted in
the application throwing an exception when viewing a page that had
missing head/office/legislature items (almost all of them).

This change fixes that by only looking for the presence of the item.
This has the nice bonus of reading better too :tada: