Disable Ruby warnings when running tests
Web integration tests introduced in f3b6b17a result in warnings like the
following being shown when running tests:

  legislative-explorer/views/layout.erb:8: warning: instance variable @head_tags not initialized

This change disables Ruby warnings when running the tests to suppress
these messages. We've done this on other EveryPolitician projects so it
seems reasonable to do it here too (even if it feels like a bit of a
hokey solution).