Don't list the same legislature repeatedly
When Wikidata has multiple values for one of the fields we're querying
the API returns multiple rows for a single country/division/city. This
was resulting in the `legislatures` method returning the same
legislature multiple times if there was another field that was causing
the API result to have multiple rows for the same country/division/city.

This change ensures each legislature for a place is only shown once.