Allow "Basic web requests" to run from new cassette
Previously, deleting the cassette and re-running the test would fall
over, as the test makes three requests, and the second and third won't
be found in the cassette generated from the first one.

I'm not sure how this came to be like this in the first place, but this
allows the cassette to be re-used.