Refactor legislatures method to be a lookup
Instead of taking a parameter and returning a place's legislatures this
now returns a hash that can be used to lookup a place's legislatures.

This is being done in the hope that this will make it easier for us to
add data to what's being returned with the legislatures. It could also
be used to cache this result, not that it's expensive, and therefore
required, right now.