Divisions and cities with multiple values for population count
Previously if a place had multiple values for the population count the
item for that place would be repeated in our list with no indication
that it was because of this reason.

Now we remove that repetition and display the multiple values for the
population count beside the name, as you would expect.