update Basic_web_requests fixture
Regenerate the `Basic_web_requests` fixture to account for the revised
SPARQL query.

Commands run:
  * rm t/fixtures/vcr/Basic_web_requests.yml
  * bundle exec rake test

(To avoid this also switching the order of episodes in the file, as the
three tests run in different orders each time, we ran the above a few
different times until the recording order stayed the same.)

This also (unexpectedly) changed the results of the HTTP fetch of "/"
within the app itself. We would have expected that to be the same as
before, as the goal of this change was simply to revert back to what we
previously had before the change of Wikidata modelling. Under most
circumstances, a change here would be a sign that our new query was
getting different results than before.

However, when we previously generated this fixture it was in the window
of time when someone had changed the name of Chile to "argentina" on
Wikidata:

  https://www.wikidata.org/w/index.php?title=Q298&type=revision&diff=635500459&oldid=635307354

We have verified (by printing out the raw HTML in the previous and new
version of this fixture) that the only diff here is in is in the _name_
of that country (and thus the ordering on the page) and the IDs of all
the countries are entirely unchanged:

  226a227,233
  >               <a class="avatar-unit" href="/country/Q298">
  >                   <span class="avatar"><i class="fa fa-users"></i></span>
  >                   <h3>Chile</h3>
  >               </a>
  >           </li>
  >
  >           <li>
  1087,1093d1093
  <           <li>
  <               <a class="avatar-unit" href="/country/Q298">
  <                   <span class="avatar"><i class="fa fa-users"></i></span>
  <                   <h3>argentina</h3>
  <               </a>
  <           </li>
  <

If we had noticed this issue earlier, we could have regenerated the
fixture independently of the current change, and this commit would only
have included changes relating to the changed SPARQL. However, it is not
possible to do that separately now as the underlying wikidata structure
has also changed.

(Tony/Alex)